refactor: co-location 적용하여 Provider 폴더 생성
