+++ src/App.tsx
@@ -1,14 +1,13 @@
 import GlobalStyle from "./style/global";
-import { router } from "./router";
-import { RouterProvider } from "react-router-dom";
-import { ModalContextProvider } from "./context";
+import { PageRouterProvider } from "./provider/router";
+import { ModalContextProvider } from "./provider/context/modal";
 
 const App = () => {
   return (
     <>
       <GlobalStyle />
       <ModalContextProvider>
-        <RouterProvider router={router} />
+        <PageRouterProvider />
       </ModalContextProvider>
     </>
   );
